Updating version to be 1.8 to stay in line of 1.x
See https://hub.docker.com/_/influxdb
latest updated to InfluxDB 2.x
The latest tag for this image now points to the latest released implementation of InfluxDB 2.x. If you are using the latest tag and would like to stay on the InfluxDB 1.x line, please update your environment to reference the 1.8 tag.